Move files into nice dirs
