+++ src/PowerShellEditorServices.Hosting/Configuration/EditorServicesConfig.cs
