+++ src/PowerShellEditorServices.Hosting/Configuration/HostInfo.cs
