+++ src/PowerShellEditorServices.Hosting/Configuration/HostLogger.cs
