+++ src/PowerShellEditorServices.Hosting/Configuration/SessionFileWriter.cs
