+++ src/PowerShellEditorServices.Hosting/Configuration/TransportConfig.cs
