+++ src/PowerShellEditorServices.Hosting/Internal/EditorServicesRunner.cs
@@ -6,10 +6,8 @@
 using Microsoft.PowerShell.EditorServices.Server;
 using System;
 using System.Collections.Generic;
-using System.Collections.ObjectModel;
 using System.IO;
 using System.Threading.Tasks;
-using SMA = System.Management.Automation;
 
 namespace Microsoft.PowerShell.EditorServices.Hosting
 {
@@ -243,14 +241,14 @@ namespace Microsoft.PowerShell.EditorServices.Hosting
             _config.PSHost.UI.WriteLine(@"
 
 
-__/\\\\\\\\\\\\\_______/\\\\\\\\\\\____/\\\\\\\\\\\________/\\\\\\\\\_        
- _\/\\\/////////\\\___/\\\/////////\\\_\/////\\\///______/\\\////////__       
-  _\/\\\_______\/\\\__\//\\\______\///______\/\\\_______/\\\/___________      
-   _\/\\\\\\\\\\\\\/____\////\\\_____________\/\\\______/\\\_____________     
-    _\/\\\/////////_________\////\\\__________\/\\\_____\/\\\_____________    
-     _\/\\\_____________________\////\\\_______\/\\\_____\//\\\____________   
-      _\/\\\______________/\\\______\//\\\______\/\\\______\///\\\__________  
-       _\/\\\_____________\///\\\\\\\\\\\/____/\\\\\\\\\\\____\////\\\\\\\\\_ 
+__/\\\\\\\\\\\\\_______/\\\\\\\\\\\____/\\\\\\\\\\\________/\\\\\\\\\_
+ _\/\\\/////////\\\___/\\\/////////\\\_\/////\\\///______/\\\////////__
+  _\/\\\_______\/\\\__\//\\\______\///______\/\\\_______/\\\/___________
+   _\/\\\\\\\\\\\\\/____\////\\\_____________\/\\\______/\\\_____________
+    _\/\\\/////////_________\////\\\__________\/\\\_____\/\\\_____________
+     _\/\\\_____________________\////\\\_______\/\\\_____\//\\\____________
+      _\/\\\______________/\\\______\//\\\______\/\\\______\///\\\__________
+       _\/\\\_____________\///\\\\\\\\\\\/____/\\\\\\\\\\\____\////\\\\\\\\\_
         _\///________________\///////////_____\///////////________\/////////__
