+++ src/PowerShellEditorServices.Hosting/Internal/NamedPipeUtils.cs
